HW-3 메인화면과 관련있는 클래스를 main 패키지로 모음.
