+++ bean628/MyProject/app/src/main/java/com/practice/achitecture/myproject/main/SearchBlogAndNewsAdapter.kt
@@ -1,21 +1,20 @@
-package com.practice.achitecture.myproject
+package com.practice.achitecture.myproject.main
 
 import android.content.Intent
+import android.graphics.Color
 import android.net.Uri
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.text.HtmlCompat
 import androidx.recyclerview.widget.RecyclerView
+import com.practice.achitecture.myproject.R
 import com.practice.achitecture.myproject.model.SearchedItem
-import common.GlideWrapper
-import kotlinx.android.synthetic.main.item_blog_and_news.view.tv_description
-import kotlinx.android.synthetic.main.item_blog_and_news.view.tv_title
-import kotlinx.android.synthetic.main.item_book_and_movie.view.*
+import kotlinx.android.synthetic.main.item_blog_and_news.view.*
 
-class SearchMovieAndBookAdapter : RecyclerView.Adapter<SearchMovieAndBookAdapter.ViewHolder>() {
+class SearchBlogAndNewsAdapter : RecyclerView.Adapter<SearchBlogAndNewsAdapter.ViewHolder>() {
 
-    private var items: ArrayList<SearchedItem> = ArrayList()
+    private var items: ArrayList<SearchedItem> = arrayListOf()
 
     override fun getItemCount() = items.size
 
@@ -36,8 +35,9 @@ class SearchMovieAndBookAdapter : RecyclerView.Adapter<SearchMovieAndBookAdapter
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val inflatedView: View = LayoutInflater.from(parent.context)
-            .inflate(R.layout.item_book_and_movie, parent, false)
-        val viewHolder = ViewHolder(inflatedView)
+            .inflate(R.layout.item_blog_and_news, parent, false)
+        val viewHolder =
+            ViewHolder(inflatedView)
         inflatedView.setOnClickListener {
             parent.context.startActivity(
                 Intent(
@@ -50,21 +50,17 @@ class SearchMovieAndBookAdapter : RecyclerView.Adapter<SearchMovieAndBookAdapter
         return viewHolder
     }
 
-    class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+    class ViewHolder(v: View) : RecyclerView.ViewHolder(v) {
 
         fun bind(item: SearchedItem?) {
-
-            if (item?.image != null) {
-                GlideWrapper.showImage(itemView.iv_main_image, item.image)
-            }
-
             val title = item?.title ?: ""
-            val content = item?.description ?: item?.director ?: ""
+            val description = item?.description ?: ""
 
+            itemView.setBackgroundColor(Color.WHITE)
             itemView.tv_title.text =
                 HtmlCompat.fromHtml(title, HtmlCompat.FROM_HTML_MODE_COMPACT)
             itemView.tv_description.text =
-                HtmlCompat.fromHtml(content, HtmlCompat.FROM_HTML_MODE_COMPACT)
+                HtmlCompat.fromHtml(description, HtmlCompat.FROM_HTML_MODE_COMPACT)
         }
     }
 }
